requirements.txt
requests==2.27.1
lxml==4.8.0
pandas==1.4.2
psycopg2==2.9.5
sqlalchemy==1.4.36
python-dotenv==0.20.0